services/inspection/requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements-dev.in
#
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
blis==1.3.3
    # via thinc
boolean-py==5.0
    # via license-expression
cachecontrol[filecache]==0.14.4
    # via
    #   cachecontrol
    #   pip-audit
catalogue==2.0.10
    # via
    #   spacy
    #   srsly
    #   thinc
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   typer-slim
    #   uvicorn
cloudpathlib==0.23.0
    # via weasel
confection==0.1.5
    # via
    #   thinc
    #   weasel
cyclonedx-python-lib==11.6.0
    # via pip-audit
cymem==2.0.13
    # via
    #   preshed
    #   spacy
    #   thinc
defusedxml==0.7.1
    # via py-serializable
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl
    # via -r requirements.in
fastapi==0.123.4
    # via -r requirements.in
filelock==3.20.0
    # via
    #   cachecontrol
    #   tldextract
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via -r requirements.in
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   tldextract
iniconfig==2.3.0
    # via pytest
jinja2==3.1.6
    # via spacy
license-expression==30.4.4
    # via cyclonedx-python-lib
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
msgpack==1.1.2
    # via cachecontrol
murmurhash==1.0.15
    # via
    #   preshed
    #   spacy
    #   thinc
numpy==2.3.5
    # via
    #   blis
    #   spacy
    #   thinc
packageurl-python==0.17.6
    # via cyclonedx-python-lib
packaging==25.0
    # via
    #   pip-audit
    #   pip-requirements-parser
    #   pytest
    #   spacy
    #   thinc
    #   weasel
phonenumbers==9.0.19
    # via presidio-analyzer
pip-api==0.0.34
    # via pip-audit
pip-audit==2.10.0
    # via -r requirements-dev.in
pip-requirements-parser==32.0.1
    # via pip-audit
platformdirs==4.5.0
    # via pip-audit
pluggy==1.6.0
    # via pytest
preshed==3.0.12
    # via
    #   spacy
    #   thinc
presidio-analyzer==2.2.359
    # via -r requirements.in
py-serializable==2.1.0
    # via cyclonedx-python-lib
pydantic==2.12.5
    # via
    #   -r requirements.in
    #   confection
    #   fastapi
    #   pydantic-settings
    #   spacy
    #   thinc
    #   weasel
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.12.0
    # via -r requirements.in
pygments==2.19.2
    # via
    #   -r requirements-dev.in
    #   pytest
    #   rich
pyparsing==3.2.5
    # via pip-requirements-parser
pytest==9.0.1
    # via -r requirements-dev.in
python-dotenv==1.2.1
    # via
    #   pydantic-settings
    #   uvicorn
pyyaml==6.0.3
    # via
    #   -r requirements.in
    #   presidio-analyzer
    #   uvicorn
regex==2025.11.3
    # via presidio-analyzer
requests==2.32.5
    # via
    #   cachecontrol
    #   pip-audit
    #   requests-file
    #   spacy
    #   tldextract
    #   weasel
requests-file==3.0.1
    # via tldextract
rich==14.2.0
    # via pip-audit
smart-open==7.5.0
    # via weasel
sortedcontainers==2.4.0
    # via cyclonedx-python-lib
spacy==3.8.11
    # via
    #   -r requirements.in
    #   presidio-analyzer
spacy-legacy==3.0.12
    # via spacy
spacy-loggers==1.0.5
    # via spacy
srsly==2.5.2
    # via
    #   confection
    #   spacy
    #   thinc
    #   weasel
starlette==0.50.0
    # via fastapi
thinc==8.3.10
    # via spacy
tldextract==5.3.0
    # via presidio-analyzer
tomli==2.3.0
    # via pip-audit
tomli-w==1.2.0
    # via pip-audit
tqdm==4.67.1
    # via spacy
typer-slim==0.20.0
    # via
    #   spacy
    #   weasel
typing-extensions==4.15.0
    # via
    #   anyio
    #   cyclonedx-python-lib
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   starlette
    #   typer-slim
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.5.0
    # via requests
uvicorn[standard]==0.38.0
    # via -r requirements.in
uvloop==0.22.1
    # via uvicorn
wasabi==1.1.3
    # via
    #   spacy
    #   thinc
    #   weasel
watchfiles==1.1.1
    # via uvicorn
weasel==0.4.3
    # via spacy
websockets==15.0.1
    # via uvicorn
wrapt==2.0.1
    # via smart-open

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
